fixed error catching on deleteProduct

+++ backend/controllers/product.controller.js
@@ -50,11 +50,15 @@ export const createProduct = async (req,res) => {
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
 
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success: false, message: "Invalid product Id"});
+
+    }
     try{
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product deleted"});
 } catch (error){
-        res.status(404).json({success: false, message: "Product not found"});
+        res.status(500).json({success: false, message: "Server Error"});
 }
 };
